refactor: improve barchart component by standardizing imports, enhancing data handling, and simplifying logic in statistics service
