+++ src/app/pages/home/home.component.html
@@ -21,7 +21,7 @@
     <div class="py-1">
       <cad-tabs [tabs]="tabs" [groupId]="'home'">
         <cad-tab [tabId]="'lapTime'" [groupId]="'home'">
-          <cad-card class="mx-2 my-2">
+          <cad-card>
             <cad-barchart
               [laps]="d.currentActivity.laps"
               [type]="'lapTime'"
